requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless==4.10.0.84
tensorflow-cpu>=2.17.0
numpy>=2.1.0
mediapipe